pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["Oskar Wyłucki <owylucki@cloudferro.com>"]
readme = "README.md"
packages = [{include = "fast_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.92.0"}
sqlalchemy = "^2.0.4"
psycopg2 = "^2.9.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
